UserService/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
SQLAlchemy==2.0.25
asyncpg==0.29.0
bcrypt==4.1.1
pyjwt==2.8.0
python-dotenv
pydantic[email]==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
grpcio==1.60.0
grpcio-tools==1.60.0

pytest
pytest-cov
httpx==0.25.2
